Monkey with shadows

+++ raytracer/include/geom/AssimpMesh.h
@@ -10,7 +10,7 @@ class AssimpMesh : public GeometricObject
 public:
     AssimpMesh(const aiScene* scene, aiMesh* mesh);
     bool Hit(const Ray& ray, double& tmin, HitRec& hr) const override;
-    bool DidShadowHit(const Ray& ray, float& t) override;
+    bool DidShadowHit(const Ray& ray, double& t) override;
 private:
     const aiScene *scene;
     aiMesh *mesh;
