+++ raytracer/include/geom/Sphere.h
@@ -10,7 +10,7 @@ public:
     bool Hit(const Ray& ray, double& tmin, HitRec& hr) const;
     void SetCenter(const glm::vec3& center);
     void SetRadius(double radius);
-    bool DidShadowHit(const Ray& ray, float& t) override;
+    bool DidShadowHit(const Ray& ray, double& t) override;
 private:
     glm::vec3 center;
     double radius;
